Lesson 1.2: Created TodoList.js and added Todo List Component. Created AddTodoForm.js and added Todo Form Component.
