+++ src/App.js
@@ -1,41 +1,17 @@
-import React from 'react';
+import React from "react";
+import AddTodoForm from "./AddTodoForm";
+import TodoList from "./TodoList";
 
-const title = 'Todo List';
-const todoList = [ 
-  {
-    title: "Complete lesson material",
-    objectID: 0,
-  },
-  {
-    title: "Complete assignment",
-    objectID: 1,
-  },
-  {
-    title: "Submit assignment",
-    objectID: 2,
-  }
-];
+const title = "Todo List";
 
-function App() {
-  return (
-    <div style={{ textAlign: 'center' }}>
-      <header>
-        <h1>
-          {title}
-        </h1>
-        </header>
-
-        <ul>
-            {todoList.map(function(item) {
-              return (
-                <li key={item.objectID}>
-                  <span>{item.title}</span>
-                </li>
-              );
-            })};
-        </ul>
-    </div>
-  );
-}
+const App = () => (
+  <div style={{ textAlign: "center" }}>
+    <header>
+      <h1>{title}</h1>
+    </header>
+    <AddTodoForm />
+    <TodoList />
+  </div>
+);
 
 export default App;
